requirements.txt
google-generativeai
tavily-python
yfinance
streamlit
pandas
requests
beautifulsoup4
pypdf
finvizfinance
colorama
